調整手機版面、拆分component
